feat: kb mode - added encoder read threshold
Added Configuration ENC_KEYCODE_READ_THRESHOLD to controller_config.h

This feature is used to reduce the limit the key repeat rate when using
the encoders to send key presses.

By default at threshold of 1, it will send keypresses at max polling
rate (1000Hz). The threshold value is basically a divisor for the
polling rate on how fast it should send a keypress with the encoders